JDBC-243 / JDBC-187 implements stream, reader and blob methods currently throwing FBDriverNotCapableException

+++ src/main/org/firebirdsql/jdbc/field/FBFlushableField.java
@@ -35,9 +35,9 @@ public interface FBFlushableField {
         public byte[] bytes;
         public InputStream binaryStream;
         public Reader characterStream;
-        public int length;
+        public long length;
 
-        public CachedObject(byte[] bytes, InputStream binaryStream, Reader characterStream, int length) {
+        public CachedObject(byte[] bytes, InputStream binaryStream, Reader characterStream, long length) {
             this.bytes = bytes;
             this.binaryStream = binaryStream;
             this.characterStream = characterStream;
